hydrate_wallets/requirements.txt
substrate-interface==1.7.9
python-dotenv==1.0.0